Refactorings: LecturerAuthenticationInteractor and StudentAuthenticationInteractor now extend superclass named UserAuthenticationInteractor
